Avanzada histórica x2

+++ database/migrations/2025_01_16_174131_create_prizes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->enum('rarity',['Común', 'Rara', 'Especial','Épica','Legendaria']);
-            $table->float('reward');
+            $table->integer('reward');
             $table->string('image');
             $table->string('audio');
             $table->timestamps();
